docs: update radio stories to include radio with labels on the left
